+++ x-pack/plugins/security_solution/public/entity_analytics/api/hooks/use_disable_risk_engine_mutation.ts
@@ -11,7 +11,7 @@ import { useInvalidateRiskEngineStatusQuery } from './use_risk_engine_status';
 import type {
   EnableRiskEngineResponse,
   EnableDisableRiskEngineErrorResponse,
-} from '../../../../server/lib/entity_analytics/risk_engine/types';
+} from '../../../../server/lib/entity_analytics/types';
 
 export const DISABLE_RISK_ENGINE_MUTATION_KEY = ['POST', 'DISABLE_RISK_ENGINE'];
