+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/risk_engine_data_client.test.ts
@@ -6,10 +6,6 @@
  */
 
 import {
-  createOrUpdateComponentTemplate,
-  createOrUpdateIndexTemplate,
-} from '@kbn/alerting-plugin/server';
-import {
   loggingSystemMock,
   elasticsearchServiceMock,
   savedObjectsClientMock,
@@ -17,12 +13,11 @@ import {
 import { taskManagerMock } from '@kbn/task-manager-plugin/server/mocks';
 import type { SavedObject } from '@kbn/core/server';
 import { RiskEngineDataClient } from './risk_engine_data_client';
-import type { RiskEngineConfiguration } from './types';
-import { createDataStream } from '../utils/create_datastream';
+import { RiskScoreDataClient } from '../risk_score/risk_score_data_client';
+import { AssetCriticalityDataClient } from '../asset_criticality/asset_criticality_data_client';
+import type { RiskEngineConfiguration } from '../types';
 import * as savedObjectConfig from './utils/saved_object_configuration';
-import * as transforms from './utils/transforms';
-import { createOrUpdateIndex } from '../utils/create_index';
-import { assetCriticalityDataClientMock } from '../asset_criticality/asset_criticality_data_client.mock';
+import * as transforms from '../utils/transforms';
 
 const getSavedObjectConfiguration = (attributes = {}) => ({
   page: 1,
@@ -48,17 +43,6 @@ const getSavedObjectConfiguration = (attributes = {}) => ({
   ],
 });
 
-const transformsMock = {
-  count: 1,
-  transforms: [
-    {
-      id: 'ml_hostriskscore_pivot_transform_default',
-      dest: { index: '' },
-      source: { index: '' },
-    },
-  ],
-};
-
 jest.mock('@kbn/alerting-plugin/server', () => ({
   createOrUpdateComponentTemplate: jest.fn(),
   createOrUpdateIndexTemplate: jest.fn(),
@@ -66,10 +50,6 @@ jest.mock('@kbn/alerting-plugin/server', () => ({
 
 jest.mock('../utils/create_datastream', () => ({
   createDataStream: jest.fn(),
-}));
-
-jest.mock('../../risk_score/transform/helpers/transforms', () => ({
-  createAndStartTransform: jest.fn(),
 }));
 
 jest.mock('../utils/create_index', () => ({
@@ -88,7 +68,6 @@ describe('RiskEngineDataClient', () => {
       let mockSavedObjectClient: ReturnType<typeof savedObjectsClientMock.create>;
       let logger: ReturnType<typeof loggingSystemMock.createLogger>;
       const esClient = elasticsearchServiceMock.createScopedClusterClient().asCurrentUser;
-      const totalFieldsLimit = 1000;
 
       beforeEach(() => {
         logger = loggingSystemMock.createLogger();
@@ -105,484 +84,6 @@ describe('RiskEngineDataClient', () => {
 
       afterEach(() => {
         jest.clearAllMocks();
-      });
-
-      describe('getWriter', () => {
-        it('should return a writer object', async () => {
-          const writer = await riskEngineDataClient.getWriter({ namespace: 'default' });
-          expect(writer).toBeDefined();
-          expect(typeof writer?.bulk).toBe('function');
-        });
-
-        it('should cache and return the same writer for the same namespace', async () => {
-          const writer1 = await riskEngineDataClient.getWriter({ namespace: 'default' });
-          const writer2 = await riskEngineDataClient.getWriter({ namespace: 'default' });
-          const writer3 = await riskEngineDataClient.getWriter({ namespace: 'space-1' });
-
-          expect(writer1).toEqual(writer2);
-          expect(writer2).not.toEqual(writer3);
-        });
-      });
-
-      describe('initializeResources success', () => {
-        it('should initialize risk engine resources', async () => {
-          await riskEngineDataClient.initializeResources({ namespace: 'default' });
-
-          expect(createOrUpdateComponentTemplate).toHaveBeenCalledWith(
-            expect.objectContaining({
-              logger,
-              esClient,
-              template: expect.objectContaining({
-                name: '.risk-score-mappings',
-                _meta: {
-                  managed: true,
-                },
-              }),
-              totalFieldsLimit: 1000,
-            })
-          );
-          expect((createOrUpdateComponentTemplate as jest.Mock).mock.lastCall[0].template.template)
-            .toMatchInlineSnapshot(`
-            Object {
-              "mappings": Object {
-                "dynamic": "strict",
-                "properties": Object {
-                  "@timestamp": Object {
-                    "ignore_malformed": false,
-                    "type": "date",
-                  },
-                  "host": Object {
-                    "properties": Object {
-                      "name": Object {
-                        "type": "keyword",
-                      },
-                      "risk": Object {
-                        "properties": Object {
-                          "calculated_level": Object {
-                            "type": "keyword",
-                          },
-                          "calculated_score": Object {
-                            "type": "float",
-                          },
-                          "calculated_score_norm": Object {
-                            "type": "float",
-                          },
-                          "category_1_count": Object {
-                            "type": "long",
-                          },
-                          "category_1_score": Object {
-                            "type": "float",
-                          },
-                          "id_field": Object {
-                            "type": "keyword",
-                          },
-                          "id_value": Object {
-                            "type": "keyword",
-                          },
-                          "inputs": Object {
-                            "properties": Object {
-                              "category": Object {
-                                "type": "keyword",
-                              },
-                              "description": Object {
-                                "type": "keyword",
-                              },
-                              "id": Object {
-                                "type": "keyword",
-                              },
-                              "index": Object {
-                                "type": "keyword",
-                              },
-                              "risk_score": Object {
-                                "type": "float",
-                              },
-                              "timestamp": Object {
-                                "type": "date",
-                              },
-                            },
-                            "type": "object",
-                          },
-                          "notes": Object {
-                            "type": "keyword",
-                          },
-                        },
-                        "type": "object",
-                      },
-                    },
-                  },
-                  "user": Object {
-                    "properties": Object {
-                      "name": Object {
-                        "type": "keyword",
-                      },
-                      "risk": Object {
-                        "properties": Object {
-                          "calculated_level": Object {
-                            "type": "keyword",
-                          },
-                          "calculated_score": Object {
-                            "type": "float",
-                          },
-                          "calculated_score_norm": Object {
-                            "type": "float",
-                          },
-                          "category_1_count": Object {
-                            "type": "long",
-                          },
-                          "category_1_score": Object {
-                            "type": "float",
-                          },
-                          "id_field": Object {
-                            "type": "keyword",
-                          },
-                          "id_value": Object {
-                            "type": "keyword",
-                          },
-                          "inputs": Object {
-                            "properties": Object {
-                              "category": Object {
-                                "type": "keyword",
-                              },
-                              "description": Object {
-                                "type": "keyword",
-                              },
-                              "id": Object {
-                                "type": "keyword",
-                              },
-                              "index": Object {
-                                "type": "keyword",
-                              },
-                              "risk_score": Object {
-                                "type": "float",
-                              },
-                              "timestamp": Object {
-                                "type": "date",
-                              },
-                            },
-                            "type": "object",
-                          },
-                          "notes": Object {
-                            "type": "keyword",
-                          },
-                        },
-                        "type": "object",
-                      },
-                    },
-                  },
-                },
-              },
-              "settings": Object {},
-            }
-          `);
-
-          expect(createOrUpdateIndexTemplate).toHaveBeenCalledWith({
-            logger,
-            esClient,
-            template: {
-              name: '.risk-score.risk-score-default-index-template',
-              body: {
-                data_stream: { hidden: true },
-                index_patterns: ['risk-score.risk-score-default'],
-                composed_of: ['.risk-score-mappings'],
-                template: {
-                  lifecycle: {},
-                  settings: {
-                    'index.mapping.total_fields.limit': totalFieldsLimit,
-                  },
-                  mappings: {
-                    dynamic: false,
-                    _meta: {
-                      kibana: {
-                        version: '8.9.0',
-                      },
-                      managed: true,
-                      namespace: 'default',
-                    },
-                  },
-                },
-                _meta: {
-                  kibana: {
-                    version: '8.9.0',
-                  },
-                  managed: true,
-                  namespace: 'default',
-                },
-              },
-            },
-          });
-
-          expect(createDataStream).toHaveBeenCalledWith({
-            logger,
-            esClient,
-            totalFieldsLimit,
-            indexPatterns: {
-              template: `.risk-score.risk-score-default-index-template`,
-              alias: `risk-score.risk-score-default`,
-            },
-          });
-
-          expect(createOrUpdateIndex).toHaveBeenCalledWith({
-            logger,
-            esClient,
-            options: {
-              index: `risk-score.risk-score-latest-default`,
-              mappings: {
-                dynamic: 'strict',
-                properties: {
-                  '@timestamp': {
-                    ignore_malformed: false,
-                    type: 'date',
-                  },
-                  host: {
-                    properties: {
-                      name: {
-                        type: 'keyword',
-                      },
-                      risk: {
-                        properties: {
-                          calculated_level: {
-                            type: 'keyword',
-                          },
-                          calculated_score: {
-                            type: 'float',
-                          },
-                          calculated_score_norm: {
-                            type: 'float',
-                          },
-                          category_1_count: {
-                            type: 'long',
-                          },
-                          category_1_score: {
-                            type: 'float',
-                          },
-                          id_field: {
-                            type: 'keyword',
-                          },
-                          id_value: {
-                            type: 'keyword',
-                          },
-                          inputs: {
-                            properties: {
-                              category: {
-                                type: 'keyword',
-                              },
-                              description: {
-                                type: 'keyword',
-                              },
-                              id: {
-                                type: 'keyword',
-                              },
-                              index: {
-                                type: 'keyword',
-                              },
-                              risk_score: {
-                                type: 'float',
-                              },
-                              timestamp: {
-                                type: 'date',
-                              },
-                            },
-                            type: 'object',
-                          },
-                          notes: {
-                            type: 'keyword',
-                          },
-                        },
-                        type: 'object',
-                      },
-                    },
-                  },
-                  user: {
-                    properties: {
-                      name: {
-                        type: 'keyword',
-                      },
-                      risk: {
-                        properties: {
-                          calculated_level: {
-                            type: 'keyword',
-                          },
-                          calculated_score: {
-                            type: 'float',
-                          },
-                          calculated_score_norm: {
-                            type: 'float',
-                          },
-                          category_1_count: {
-                            type: 'long',
-                          },
-                          category_1_score: {
-                            type: 'float',
-                          },
-                          id_field: {
-                            type: 'keyword',
-                          },
-                          id_value: {
-                            type: 'keyword',
-                          },
-                          inputs: {
-                            properties: {
-                              category: {
-                                type: 'keyword',
-                              },
-                              description: {
-                                type: 'keyword',
-                              },
-                              id: {
-                                type: 'keyword',
-                              },
-                              index: {
-                                type: 'keyword',
-                              },
-                              risk_score: {
-                                type: 'float',
-                              },
-                              timestamp: {
-                                type: 'date',
-                              },
-                            },
-                            type: 'object',
-                          },
-                          notes: {
-                            type: 'keyword',
-                          },
-                        },
-                        type: 'object',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          });
-
-          expect(transforms.createTransform).toHaveBeenCalledWith({
-            logger,
-            esClient,
-            transform: {
-              dest: {
-                index: 'risk-score.risk-score-latest-default',
-              },
-              frequency: '1h',
-              latest: {
-                sort: '@timestamp',
-                unique_key: ['host.name', 'user.name'],
-              },
-              source: {
-                index: ['risk-score.risk-score-default'],
-              },
-              sync: {
-                time: {
-                  delay: '2s',
-                  field: '@timestamp',
-                },
-              },
-              transform_id: 'risk_score_latest_transform_default',
-            },
-          });
-        });
-      });
-
-      describe('initializeResources error', () => {
-        it('should handle errors during initialization', async () => {
-          const error = new Error('There error');
-          (createOrUpdateIndexTemplate as jest.Mock).mockRejectedValueOnce(error);
-
-          try {
-            await riskEngineDataClient.initializeResources({ namespace: 'default' });
-          } catch (e) {
-            expect(logger.error).toHaveBeenCalledWith(
-              `Error initializing risk engine resources: ${error.message}`
-            );
-          }
-        });
-      });
-
-      describe('getStatus', () => {
-        it('should return initial status', async () => {
-          const status = await riskEngineDataClient.getStatus({
-            namespace: 'default',
-          });
-          expect(status).toEqual({
-            isMaxAmountOfRiskEnginesReached: false,
-            riskEngineStatus: 'NOT_INSTALLED',
-            legacyRiskEngineStatus: 'NOT_INSTALLED',
-          });
-        });
-
-        describe('saved object exists and transforms not', () => {
-          beforeEach(() => {
-            mockSavedObjectClient.find.mockResolvedValue(getSavedObjectConfiguration());
-          });
-
-          it('should return status with enabled true', async () => {
-            mockSavedObjectClient.find.mockResolvedValue(
-              getSavedObjectConfiguration({
-                enabled: true,
-              })
-            );
-
-            const status = await riskEngineDataClient.getStatus({
-              namespace: 'default',
-            });
-            expect(status).toEqual({
-              isMaxAmountOfRiskEnginesReached: true,
-              riskEngineStatus: 'ENABLED',
-              legacyRiskEngineStatus: 'NOT_INSTALLED',
-            });
-          });
-
-          it('should return status with enabled false', async () => {
-            mockSavedObjectClient.find.mockResolvedValue(getSavedObjectConfiguration());
-
-            const status = await riskEngineDataClient.getStatus({
-              namespace: 'default',
-            });
-            expect(status).toEqual({
-              isMaxAmountOfRiskEnginesReached: false,
-              riskEngineStatus: 'DISABLED',
-              legacyRiskEngineStatus: 'NOT_INSTALLED',
-            });
-          });
-        });
-
-        describe('legacy transforms', () => {
-          it('should fetch transforms', async () => {
-            await riskEngineDataClient.getStatus({
-              namespace: 'default',
-            });
-
-            expect(esClient.transform.getTransform).toHaveBeenCalledTimes(4);
-            expect(esClient.transform.getTransform).toHaveBeenNthCalledWith(1, {
-              transform_id: 'ml_hostriskscore_pivot_transform_default',
-            });
-            expect(esClient.transform.getTransform).toHaveBeenNthCalledWith(2, {
-              transform_id: 'ml_hostriskscore_latest_transform_default',
-            });
-            expect(esClient.transform.getTransform).toHaveBeenNthCalledWith(3, {
-              transform_id: 'ml_userriskscore_pivot_transform_default',
-            });
-            expect(esClient.transform.getTransform).toHaveBeenNthCalledWith(4, {
-              transform_id: 'ml_userriskscore_latest_transform_default',
-            });
-          });
-
-          it('should return that legacy transform enabled if at least on transform exist', async () => {
-            esClient.transform.getTransform.mockResolvedValueOnce(transformsMock);
-
-            const status = await riskEngineDataClient.getStatus({
-              namespace: 'default',
-            });
-
-            expect(status).toEqual({
-              isMaxAmountOfRiskEnginesReached: false,
-              riskEngineStatus: 'NOT_INSTALLED',
-              legacyRiskEngineStatus: 'ENABLED',
-            });
-
-            esClient.transform.getTransformStats.mockReset();
-          });
-        });
       });
 
       describe('#getConfiguration', () => {
@@ -704,10 +205,8 @@ describe('RiskEngineDataClient', () => {
 
       describe('init', () => {
         let mockTaskManagerStart: ReturnType<typeof taskManagerMock.createStart>;
-        const initializeResourcesMock = jest.spyOn(
-          RiskEngineDataClient.prototype,
-          'initializeResources'
-        );
+        const initRiskScore = jest.spyOn(RiskScoreDataClient.prototype, 'init');
+        const initAssetCriticality = jest.spyOn(AssetCriticalityDataClient.prototype, 'init');
         const enableRiskEngineMock = jest.spyOn(RiskEngineDataClient.prototype, 'enableRiskEngine');
 
         const disableLegacyRiskEngineMock = jest.spyOn(
@@ -718,7 +217,7 @@ describe('RiskEngineDataClient', () => {
           mockTaskManagerStart = taskManagerMock.createStart();
           disableLegacyRiskEngineMock.mockImplementation(() => Promise.resolve(true));
 
-          initializeResourcesMock.mockImplementation(() => {
+          initRiskScore.mockImplementation(() => {
             return Promise.resolve();
           });
 
@@ -732,7 +231,8 @@ describe('RiskEngineDataClient', () => {
         });
 
         afterEach(() => {
-          initializeResourcesMock.mockReset();
+          initRiskScore.mockReset();
+          initAssetCriticality.mockReset();
           enableRiskEngineMock.mockReset();
           disableLegacyRiskEngineMock.mockReset();
         });
@@ -741,7 +241,6 @@ describe('RiskEngineDataClient', () => {
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
 
@@ -762,7 +261,6 @@ describe('RiskEngineDataClient', () => {
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
 
@@ -784,7 +282,6 @@ describe('RiskEngineDataClient', () => {
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
 
@@ -799,19 +296,18 @@ describe('RiskEngineDataClient', () => {
         });
 
         it('should catch error for initializeResources and stop', async () => {
-          initializeResourcesMock.mockImplementationOnce(() => {
-            throw new Error('Error initializeResourcesMock');
+          initRiskScore.mockImplementationOnce(() => {
+            throw new Error('Error initRiskScore');
           });
 
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
 
           expect(initResult).toEqual({
-            errors: ['Error initializeResourcesMock'],
+            errors: ['Error initRiskScore'],
             legacyRiskEngineDisabled: true,
             riskEngineConfigurationCreated: false,
             riskEngineEnabled: false,
@@ -821,20 +317,18 @@ describe('RiskEngineDataClient', () => {
         });
 
         it('should catch error for asset criticality init and stop', async () => {
-          const assetCriticalityDataClient = assetCriticalityDataClientMock.create();
-          assetCriticalityDataClient.init.mockImplementationOnce(() => {
-            throw new Error('Error assetCriticalityDataClient');
+          initAssetCriticality.mockImplementationOnce(() => {
+            throw new Error('Error initAssetCriticality');
           });
 
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient,
             isAssetCriticalityEnabled: true,
           });
 
           expect(initResult).toEqual({
-            errors: ['Error assetCriticalityDataClient'],
+            errors: ['Error initAssetCriticality'],
             legacyRiskEngineDisabled: true,
             riskEngineConfigurationCreated: false,
             riskEngineEnabled: false,
@@ -851,7 +345,6 @@ describe('RiskEngineDataClient', () => {
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
 
@@ -873,7 +366,6 @@ describe('RiskEngineDataClient', () => {
           const initResult = await riskEngineDataClient.init({
             namespace: 'default',
             taskManager: mockTaskManagerStart,
-            assetCriticalityDataClient: assetCriticalityDataClientMock.create(),
             isAssetCriticalityEnabled: true,
           });
