+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/risk_engine_data_client.ts
@@ -5,61 +5,30 @@
  * 2.0.
  */
 
-import type { Metadata } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
-import type { ClusterPutComponentTemplateRequest } from '@elastic/elasticsearch/lib/api/types';
-import {
-  createOrUpdateComponentTemplate,
-  createOrUpdateIndexTemplate,
-} from '@kbn/alerting-plugin/server';
-import { mappingFromFieldMap } from '@kbn/alerting-plugin/common';
-import { DEFAULT_NAMESPACE_STRING } from '@kbn/core-saved-objects-utils-server';
 import type { Logger, ElasticsearchClient, SavedObjectsClientContract } from '@kbn/core/server';
 import type { TaskManagerStartContract } from '@kbn/task-manager-plugin/server';
-
-import {
-  riskScoreFieldMap,
-  getIndexPatternDataStream,
-  totalFieldsLimit,
-  mappingComponentName,
-  getTransformOptions,
-} from './configurations';
-import { createDataStream } from '../utils/create_datastream';
-import type { RiskEngineDataWriter as Writer } from './risk_engine_data_writer';
-import { RiskEngineDataWriter } from './risk_engine_data_writer';
 import type { InitRiskEngineResult } from '../../../../common/risk_engine';
 import {
   RiskEngineStatus,
-  getRiskScoreLatestIndex,
   MAX_SPACES_COUNT,
   RiskScoreEntity,
 } from '../../../../common/risk_engine';
-import {
-  getLegacyTransforms,
-  getLatestTransformId,
-  removeLegacyTransforms,
-  createTransform,
-} from './utils/transforms';
+import { removeLegacyTransforms, getLegacyTransforms } from '../utils/transforms';
 import {
   updateSavedObjectAttribute,
   getConfiguration,
   initSavedObjects,
   getEnabledRiskEngineAmount,
 } from './utils/saved_object_configuration';
-import { getRiskInputsIndex } from './get_risk_inputs_index';
-import { removeRiskScoringTask, startRiskScoringTask } from './tasks';
-import { createOrUpdateIndex } from '../utils/create_index';
 import { bulkDeleteSavedObjects } from '../../risk_score/prebuilt_saved_objects/helpers/bulk_delete_saved_objects';
-import type { AssetCriticalityDataClient } from '../asset_criticality/asset_criticality_data_client';
+import { RiskScoreDataClient } from '../risk_score/risk_score_data_client';
+import { removeRiskScoringTask, startRiskScoringTask } from '../risk_score/tasks';
+import { AssetCriticalityDataClient } from '../asset_criticality/asset_criticality_data_client';
 
 interface InitOpts {
   namespace: string;
   taskManager: TaskManagerStartContract;
-  assetCriticalityDataClient: AssetCriticalityDataClient;
   isAssetCriticalityEnabled: boolean;
-}
-
-interface InitializeRiskEngineResourcesOpts {
-  namespace?: string;
 }
 
 interface RiskEngineDataClientOpts {
@@ -71,13 +40,12 @@ interface RiskEngineDataClientOpts {
 }
 
 export class RiskEngineDataClient {
-  private writerCache: Map<string, Writer> = new Map();
   constructor(private readonly options: RiskEngineDataClientOpts) {}
 
   public async init({
     namespace,
     taskManager,
-    assetCriticalityDataClient,
+
     isAssetCriticalityEnabled,
   }: InitOpts) {
     const result: InitRiskEngineResult = {
@@ -97,7 +65,14 @@ export class RiskEngineDataClient {
     }
 
     try {
-      await this.initializeResources({ namespace });
+      const riskScoreDataClient = new RiskScoreDataClient({
+        esClient: this.options.esClient,
+        logger: this.options.logger,
+        namespace,
+        kibanaVersion: this.options.kibanaVersion,
+        soClient: this.options.soClient,
+      });
+      await riskScoreDataClient.init();
       result.riskEngineResourcesInstalled = true;
     } catch (e) {
       result.errors.push(e.message);
@@ -106,6 +81,12 @@ export class RiskEngineDataClient {
 
     if (isAssetCriticalityEnabled) {
       try {
+        // We use internal user here for resource installation, not current user
+        const assetCriticalityDataClient = new AssetCriticalityDataClient({
+          esClient: this.options.esClient,
+          logger: this.options.logger,
+          namespace,
+        });
         await assetCriticalityDataClient.init();
         result.assetCriticalityInstalled = true;
       } catch (e) {
@@ -137,37 +118,9 @@ export class RiskEngineDataClient {
     return result;
   }
 
-  public async getWriter({ namespace }: { namespace: string }): Promise<Writer> {
-    if (this.writerCache.get(namespace)) {
-      return this.writerCache.get(namespace) as Writer;
-    }
-    const indexPatterns = getIndexPatternDataStream(namespace);
-    await this.initializeWriter(namespace, indexPatterns.alias);
-    return this.writerCache.get(namespace) as Writer;
-  }
-
-  private async initializeWriter(namespace: string, index: string): Promise<Writer> {
-    const writer = new RiskEngineDataWriter({
-      esClient: this.options.esClient,
-      namespace,
-      index,
-      logger: this.options.logger,
-    });
-
-    this.writerCache.set(namespace, writer);
-    return writer;
-  }
-
   public getConfiguration = () =>
     getConfiguration({
       savedObjectsClient: this.options.soClient,
-    });
-
-  public getRiskInputsIndex = ({ dataViewId }: { dataViewId: string }) =>
-    getRiskInputsIndex({
-      dataViewId,
-      logger: this.options.logger,
-      soClient: this.options.soClient,
     });
 
   public async getStatus({ namespace }: { namespace: string }) {
@@ -277,97 +230,5 @@ export class RiskEngineDataClient {
     }
 
     return RiskEngineStatus.ENABLED;
-  }
-
-  public async initializeResources({
-    namespace = DEFAULT_NAMESPACE_STRING,
-  }: InitializeRiskEngineResourcesOpts) {
-    try {
-      const esClient = this.options.esClient;
-
-      const indexPatterns = getIndexPatternDataStream(namespace);
-
-      const indexMetadata: Metadata = {
-        kibana: {
-          version: this.options.kibanaVersion,
-        },
-        managed: true,
-        namespace,
-      };
-
-      await Promise.all([
-        createOrUpdateComponentTemplate({
-          logger: this.options.logger,
-          esClient,
-          template: {
-            name: mappingComponentName,
-            _meta: {
-              managed: true,
-            },
-            template: {
-              settings: {},
-              mappings: mappingFromFieldMap(riskScoreFieldMap, 'strict'),
-            },
-          } as ClusterPutComponentTemplateRequest,
-          totalFieldsLimit,
-        }),
-      ]);
-
-      await createOrUpdateIndexTemplate({
-        logger: this.options.logger,
-        esClient,
-        template: {
-          name: indexPatterns.template,
-          body: {
-            data_stream: { hidden: true },
-            index_patterns: [indexPatterns.alias],
-            composed_of: [mappingComponentName],
-            template: {
-              lifecycle: {},
-              settings: {
-                'index.mapping.total_fields.limit': totalFieldsLimit,
-              },
-              mappings: {
-                dynamic: false,
-                _meta: indexMetadata,
-              },
-            },
-            _meta: indexMetadata,
-          },
-        },
-      });
-
-      await createDataStream({
-        logger: this.options.logger,
-        esClient,
-        totalFieldsLimit,
-        indexPatterns,
-      });
-
-      await createOrUpdateIndex({
-        esClient,
-        logger: this.options.logger,
-        options: {
-          index: getRiskScoreLatestIndex(namespace),
-          mappings: mappingFromFieldMap(riskScoreFieldMap, 'strict'),
-        },
-      });
-
-      const transformId = getLatestTransformId(namespace);
-      await createTransform({
-        esClient,
-        logger: this.options.logger,
-        transform: {
-          transform_id: transformId,
-          ...getTransformOptions({
-            dest: getRiskScoreLatestIndex(namespace),
-            source: [indexPatterns.alias],
-          }),
-        },
-      });
-    } catch (error) {
-      this.options.logger.error(`Error initializing risk engine resources: ${error.message}`);
-      throw error;
-    }
   }
 }
