+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/utils/saved_object_configuration.ts
@@ -7,7 +7,7 @@
 import type { SavedObject, SavedObjectsClientContract } from '@kbn/core/server';
 
 import { getAlertsIndex } from '../../../../../common/utils/risk_score_modules';
-import type { RiskEngineConfiguration } from '../types';
+import type { RiskEngineConfiguration } from '../../types';
 import { riskEngineConfigurationTypeName } from '../saved_object';
 
 export interface SavedObjectsClientArg {
