+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/calculate_and_persist_risk_scores.mock.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { CalculateAndPersistScoresResponse } from './types';
+import type { CalculateAndPersistScoresResponse } from '../types';
 
 const buildResponseMock = (
   overrides: Partial<CalculateAndPersistScoresResponse> = {}
