+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/calculate_risk_scores.mock.ts
@@ -14,7 +14,7 @@ import type {
   CalculateRiskScoreAggregations,
   CalculateScoresResponse,
   RiskScoreBucket,
-} from './types';
+} from '../types';
 
 const buildRiskScoreBucketMock = (overrides: Partial<RiskScoreBucket> = {}): RiskScoreBucket => ({
   key: { 'user.name': 'username' },
