+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/calculate_risk_scores.ts
@@ -38,7 +38,7 @@ import type {
   CalculateScoresParams,
   CalculateScoresResponse,
   RiskScoreBucket,
-} from './types';
+} from '../types';
 
 const bucketToResponse = ({
   bucket,
