+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/configurations.ts
