+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/get_risk_inputs_index.mock.ts
