+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/helpers.test.ts
