+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/risk_score_data_client.mock.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { RiskScoreDataClient } from './risk_score_data_client';
+
+const createRiskScoreDataClientMock = () =>
+  ({
+    getWriter: jest.fn(),
+    init: jest.fn(),
+    getRiskInputsIndex: jest.fn(),
+  } as unknown as jest.Mocked<RiskScoreDataClient>);
+
+export const riskScoreDataClientMock = { create: createRiskScoreDataClientMock };
