+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/risk_score_service.mock.ts
