+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/tasks/helpers.test.ts
