+++ x-pack/plugins/security_solution/server/lib/entity_analytics/routes/risk_engine_disable_route.test.ts
@@ -8,13 +8,13 @@
 import { taskManagerMock } from '@kbn/task-manager-plugin/server/mocks';
 import { riskEngineDisableRoute } from './risk_engine_disable_route';
 
-import { RISK_ENGINE_DISABLE_URL } from '../../../../../common/constants';
+import { RISK_ENGINE_DISABLE_URL } from '../../../../common/constants';
 import {
   serverMock,
   requestContextMock,
   requestMock,
-} from '../../../detection_engine/routes/__mocks__';
-import { riskEngineDataClientMock } from '../risk_engine_data_client.mock';
+} from '../../detection_engine/routes/__mocks__';
+import { riskEngineDataClientMock } from '../risk_engine/risk_engine_data_client.mock';
 
 describe('risk score disable route', () => {
   let server: ReturnType<typeof serverMock.create>;
