+++ x-pack/plugins/security_solution/server/lib/entity_analytics/routes/risk_engine_privileges_route.ts
