+++ x-pack/plugins/security_solution/server/lib/entity_analytics/routes/translations.ts
