+++ x-pack/plugins/security_solution/server/types.ts
@@ -29,7 +29,7 @@ import type {
 import type { FrameworkRequest } from './lib/framework';
 import type { EndpointAuthz } from '../common/endpoint/types/authz';
 import type { EndpointInternalFleetServicesInterface } from './endpoint/services/fleet';
-import type { RiskEngineDataClient } from './lib/entity_analytics/risk_engine/risk_engine_data_client';
+import type { RiskEngineDataClient } from './lib/entity_analytics/risk_score/risk_engine_data_client';
 
 export { AppClient };
